requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
pypdf>=3.17.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
psycopg2-binary>=2.9.9
pgvector>=0.2.3
langchain-core>=0.1.10